adding a function toextract batch_uid from json and modifying the get_batch interface for easier call

+++ crates/meilisearch/tests/common/index.rs
@@ -379,7 +379,7 @@ impl<State> Index<'_, State> {
         self.service.get(url).await
     }
 
-    pub async fn get_batch(&self, batch_id: u32) -> (Value, StatusCode) {
+    pub async fn get_batch(&self, batch_id: u64) -> (Value, StatusCode) {
         let url = format!("/batches/{}", batch_id);
         self.service.get(url).await
     }
